+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.3.4
 
 - [helpers] expand default `required` validation to include booleans
+- [slider] change `didEnter`, `didEnterAsPrev`, and `didEnterAsNext` prop types from booleans to functions 
+- [slider] fix typo in propType `didEntereAsNext`
 
 ## 0.3.3